upgradum prideleny kategorie a rarity mensi zmeny v inventari level up system a odemykani slotu v inventari

+++ Assets/scripts/Upgrade/AbstractUpgrade.cs
@@ -160,7 +160,7 @@ namespace Assets.scripts.Upgrade
 
 		public float AddValueByLevel(float baseDamage, float levelMultiplier, bool alsoNegative=false)
 		{
-			int add = (int)((Level - 1) * levelMultiplier);
+			float add = (Level - 1)*levelMultiplier;
 			if (add < 0 && !alsoNegative)
 				add = 0;
 			return (baseDamage + add);
